wip BEO-8607 change event design

+++ Core/ApiObjects/EventBus/CampaignTriggerEvent.cs
@@ -5,7 +5,7 @@ namespace ApiObjects.EventBus
     public class CampaignTriggerEvent : ServiceEvent
     {
         public long CampaignId { get; set; }
-        public int DomainId { get; set; }
+        public long DomainId { get; set; }
         public CoreObject EventObject { get; set; }
     }
 }
